Modularize hero profile hook
